[ADD: print floats in exp form]

+++ ft_divmod.c
@@ -1,6 +1,6 @@
 #include "ft_printf.h"
 
-intmax_t	ft_divmod(intmax_t dividend, intmax_t divisor, intmax_t *remainder)
+int32_t	ft_divmod(int32_t dividend, int32_t divisor, int32_t *remainder)
 {
 	*remainder = dividend % divisor;
 	return (dividend / divisor);
